Fixing issues for BSD systems

+++ sysutil_bsd.go
@@ -21,11 +21,11 @@ func init() {
 
 	// decode
 	var t syscall.Timeval32
-	err = binary.Read(bytes.NewBuffer([]byte(res), binary.LittleEndian), &t)
+	err = binary.Read(bytes.NewBuffer([]byte(res)), binary.LittleEndian, &t)
 	if err != nil {
 		btime = time.Now()
 		return
 	}
 
-	btime = time.Unix(t.Sec, t.Usec)
+	btime = time.Unix(int64(t.Sec), int64(t.Usec))
 }
